+++ CHANGELOG.md
@@ -12,6 +12,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- hnefatafl-client: maximum and minimum rating for archived games.
+
 ## 5.4.2 - 2026-03-22
 
 ### Changed